remove experiments
