+++ docs/index.html
@@ -128,18 +128,13 @@ print(model) # print the rule-based model
 </tr>
 <tr>
 <td align="left">C4.5 rule tree</td>
-<td><a href="https://csinva.io/imodels/tree/cart_wrapper.html">🗂️</a>, <a href="https://scikit-learn.org/stable/modules/tree.html">🔗</a>, <a href="https://www.taylorfrancis.com/books/mono/10.1201/9781315139470/classification-regression-trees-leo-breiman-jerome-friedman-richard-olshen-charles-stone">📄</a></td>
-<td>Greedily learns tree using CART</td>
+<td><a href="https://csinva.io/imodels/tree/c45_tree/c45_tree.html#imodels.tree.c45_tree.c45_tree.C45TreeClassifier">🗂️</a>, <a href="https://github.com/RaczeQ/scikit-learn-C4.5-tree-classifier">🔗</a>, <a href="https://link.springer.com/article/10.1007/BF00993309">📄</a></td>
+<td>Greedily learns tree using C4.5</td>
 </tr>
 <tr>
-<td align="left"></td>
-<td>Iterative random forest</td>
-<td>[🗂️]</td>
-</tr>
-<tr>
-<td align="left">(<a href="https://csinva.io/imodels/tree/iterative_random_forest/iterative_random_forest.html">https://csinva.io/imodels/tree/iterative_random_forest/iterative_random_forest.html</a>), <a href="https://github.com/Yu-Group/iterative-Random-Forest">🔗</a>, <a href="https://www.pnas.org/content/115/8/1943">📄</a></td>
+<td align="left">Iterative random forest</td>
+<td><a href="https://csinva.io/imodels/tree/iterative_random_forest/iterative_random_forest.html">🗂️</a>, <a href="https://github.com/Yu-Group/iterative-Random-Forest">🔗</a>, <a href="https://www.pnas.org/content/115/8/1943">📄</a></td>
 <td>(In progress) Repeatedly fit random forest, giving features with high importance a higher chance of being selected</td>
-<td></td>
 </tr>
 <tr>
 <td align="left">Sparse integer linear model</td>
@@ -413,6 +408,7 @@ import os
 from .tree.cart_wrapper import GreedyTreeClassifier, GreedyTreeRegressor
 from .tree.gosdt.pygosdt import OptimalTreeClassifier
 from .tree.c45_tree.c45_tree import C45TreeClassifier
+from .tree.saps import SaplingSumRegressor, SaplingSumClassifier
 from .algebraic.slim import SLIMRegressor, SLIMClassifier
 from .discretization.discretizer import RFDiscretizer, BasicDiscretizer
 from .discretization.mdlp import MDLPDiscretizer, BRLDiscretizer
@@ -433,8 +429,11 @@ from .rule_set.slipper import SlipperClassifier
 from .util.explain_errors import explain_classification_errors
 
 CLASSIFIERS = [BayesianRuleListClassifier, GreedyRuleListClassifier, SkopeRulesClassifier,
-               BoostedRulesClassifier, SLIMClassifier, SlipperClassifier, BayesianRuleSetClassifier]  # , IRFClassifier
-REGRESSORS = [RuleFitRegressor, SLIMRegressor]
+               BoostedRulesClassifier, SLIMClassifier, SlipperClassifier, BayesianRuleSetClassifier,
+               C45TreeClassifier, OptimalTreeClassifier, OptimalRuleListClassifier, OneRClassifier,
+               SlipperClassifier,
+               SaplingSumClassifier]  # , IRFClassifier
+REGRESSORS = [RuleFitRegressor, SLIMRegressor, GreedyTreeClassifier, SaplingSumRegressor]
 DISCRETIZERS = [RFDiscretizer, BasicDiscretizer, MDLPDiscretizer, BRLDiscretizer]</code></pre>
 </details>
 </section>
