+++ docs/tree/c45_tree/c45_tree.html
@@ -29,11 +29,7 @@ References
 References</p>
 <hr>
 <p>.. [1] <a href="https://en.wikipedia.org/wiki/Decision_tree_learning">https://en.wikipedia.org/wiki/Decision_tree_learning</a>
-.. [2] <a href="https://en.wikipedia.org/wiki/C4.5_algorithm">https://en.wikipedia.org/wiki/C4.5_algorithm</a>
-.. [3] L. Breiman, J. Friedman, R. Olshen, and C. Stone, "Classification
-and Regression Trees", Wadsworth, Belmont, CA, 1984.
-.. [4] J. R. Quinlain, "C4.5: Programs for Machine Learning",
-Morgan Kaufmann Publishers, 1993</p>
+.. [2] <a href="https://en.wikipedia.org/wiki/C4.5_algorithm">https://en.wikipedia.org/wiki/C4.5_algorithm</a></p>
 <details class="source">
 <summary>
 <span>Expand source code</span>
@@ -43,10 +39,6 @@ References
 ----------
 .. [1] https://en.wikipedia.org/wiki/Decision_tree_learning
 .. [2] https://en.wikipedia.org/wiki/C4.5_algorithm
-.. [3] L. Breiman, J. Friedman, R. Olshen, and C. Stone, &#34;Classification
-       and Regression Trees&#34;, Wadsworth, Belmont, CA, 1984.
-.. [4] J. R. Quinlain, &#34;C4.5: Programs for Machine Learning&#34;,
-       Morgan Kaufmann Publishers, 1993
 &#34;&#34;&#34;
 from typing import List
 from xml.dom import minidom
@@ -79,7 +71,12 @@ class C45TreeClassifier(BaseEstimator, ClassifierMixin):
         if feature_names is None:
             self.feature_names = [f&#39;X_{x}&#39; for x in range(X.shape[1])]
         else:
-            self.feature_names = feature_names
+            # only include alphanumeric chars / replace spaces with underscores
+            self.feature_names = [&#39;&#39;.join([i for i in x if i.isalnum()]).replace(&#39; &#39;, &#39;_&#39;)
+                                  for x in feature_names]
+            self.feature_names = [&#39;X_&#39; + x if x[0].isdigit()
+                                  else x
+                                  for x in feature_names]
 
         assert len(self.feature_names) == X.shape[1]
 
@@ -93,13 +90,14 @@ class C45TreeClassifier(BaseEstimator, ClassifierMixin):
         root = ET.Element(&#39;GreedyTree&#39;)
         self.grow_tree(data, categories, root, self.feature_names)  # adds to root
         self.tree_ = ET.tostring(root, encoding=&#34;unicode&#34;)
+        # print(&#39;self.tree_&#39;, self.tree_)
+        self.dom_ = minidom.parseString(self.tree_)
         return self
 
     def raw_preds(self, X):
         check_is_fitted(self, [&#39;tree_&#39;, &#39;resultType&#39;, &#39;feature_names&#39;])
         X = check_array(X)
-        dom = minidom.parseString(self.tree_)
-        root = dom.childNodes[0]
+        root = self.dom_.childNodes[0]
         prediction = []
         for i in range(X.shape[0]):
             answerlist = decision(root, X[i], self.feature_names, 1)
@@ -117,8 +115,7 @@ class C45TreeClassifier(BaseEstimator, ClassifierMixin):
 
     def __str__(self):
         check_is_fitted(self, [&#39;tree_&#39;])
-        dom = minidom.parseString(self.tree_)
-        return dom.toprettyxml(newl=&#34;\r\n&#34;)
+        return self.dom_.toprettyxml(newl=&#34;\r\n&#34;)
 
     def grow_tree(self, X_t: List[list], y_str: List[str], parent, attrs_names):
         &#34;&#34;&#34;
@@ -279,7 +276,12 @@ if __name__ == &#39;__main__&#39;:
         if feature_names is None:
             self.feature_names = [f&#39;X_{x}&#39; for x in range(X.shape[1])]
         else:
-            self.feature_names = feature_names
+            # only include alphanumeric chars / replace spaces with underscores
+            self.feature_names = [&#39;&#39;.join([i for i in x if i.isalnum()]).replace(&#39; &#39;, &#39;_&#39;)
+                                  for x in feature_names]
+            self.feature_names = [&#39;X_&#39; + x if x[0].isdigit()
+                                  else x
+                                  for x in feature_names]
 
         assert len(self.feature_names) == X.shape[1]
 
@@ -293,13 +295,14 @@ if __name__ == &#39;__main__&#39;:
         root = ET.Element(&#39;GreedyTree&#39;)
         self.grow_tree(data, categories, root, self.feature_names)  # adds to root
         self.tree_ = ET.tostring(root, encoding=&#34;unicode&#34;)
+        # print(&#39;self.tree_&#39;, self.tree_)
+        self.dom_ = minidom.parseString(self.tree_)
         return self
 
     def raw_preds(self, X):
         check_is_fitted(self, [&#39;tree_&#39;, &#39;resultType&#39;, &#39;feature_names&#39;])
         X = check_array(X)
-        dom = minidom.parseString(self.tree_)
-        root = dom.childNodes[0]
+        root = self.dom_.childNodes[0]
         prediction = []
         for i in range(X.shape[0]):
             answerlist = decision(root, X[i], self.feature_names, 1)
@@ -317,8 +320,7 @@ if __name__ == &#39;__main__&#39;:
 
     def __str__(self):
         check_is_fitted(self, [&#39;tree_&#39;])
-        dom = minidom.parseString(self.tree_)
-        return dom.toprettyxml(newl=&#34;\r\n&#34;)
+        return self.dom_.toprettyxml(newl=&#34;\r\n&#34;)
 
     def grow_tree(self, X_t: List[list], y_str: List[str], parent, attrs_names):
         &#34;&#34;&#34;
@@ -444,7 +446,12 @@ if __name__ == &#39;__main__&#39;:
     if feature_names is None:
         self.feature_names = [f&#39;X_{x}&#39; for x in range(X.shape[1])]
     else:
-        self.feature_names = feature_names
+        # only include alphanumeric chars / replace spaces with underscores
+        self.feature_names = [&#39;&#39;.join([i for i in x if i.isalnum()]).replace(&#39; &#39;, &#39;_&#39;)
+                              for x in feature_names]
+        self.feature_names = [&#39;X_&#39; + x if x[0].isdigit()
+                              else x
+                              for x in feature_names]
 
     assert len(self.feature_names) == X.shape[1]
 
@@ -458,6 +465,8 @@ if __name__ == &#39;__main__&#39;:
     root = ET.Element(&#39;GreedyTree&#39;)
     self.grow_tree(data, categories, root, self.feature_names)  # adds to root
     self.tree_ = ET.tostring(root, encoding=&#34;unicode&#34;)
+    # print(&#39;self.tree_&#39;, self.tree_)
+    self.dom_ = minidom.parseString(self.tree_)
     return self</code></pre>
 </details>
 </dd>
@@ -621,8 +630,7 @@ if __name__ == &#39;__main__&#39;:
 <pre><code class="python">def raw_preds(self, X):
     check_is_fitted(self, [&#39;tree_&#39;, &#39;resultType&#39;, &#39;feature_names&#39;])
     X = check_array(X)
-    dom = minidom.parseString(self.tree_)
-    root = dom.childNodes[0]
+    root = self.dom_.childNodes[0]
     prediction = []
     for i in range(X.shape[0]):
         answerlist = decision(root, X[i], self.feature_names, 1)
