+++ docs/tree/index.html
@@ -50,6 +50,10 @@
 <dd>
 <div class="desc"><p>Repeatedly fit random forest, giving features with high importance a higher chance of being selected.</p></div>
 </dd>
+<dt><code class="name"><a title="imodels.tree.saps" href="saps.html">imodels.tree.saps</a></code></dt>
+<dd>
+<div class="desc"></div>
+</dd>
 </dl>
 </section>
 <section>
@@ -76,6 +80,7 @@
 <li><code><a title="imodels.tree.cart_wrapper" href="cart_wrapper.html">imodels.tree.cart_wrapper</a></code></li>
 <li><code><a title="imodels.tree.gosdt" href="gosdt/index.html">imodels.tree.gosdt</a></code></li>
 <li><code><a title="imodels.tree.iterative_random_forest" href="iterative_random_forest/index.html">imodels.tree.iterative_random_forest</a></code></li>
+<li><code><a title="imodels.tree.saps" href="saps.html">imodels.tree.saps</a></code></li>
 </ul>
 </li>
 </ul>
