+++ docs/util/index.html
@@ -38,6 +38,10 @@
 <dd>
 <div class="desc"></div>
 </dd>
+<dt><code class="name"><a title="imodels.util.data_util" href="data_util.html">imodels.util.data_util</a></code></dt>
+<dd>
+<div class="desc"></div>
+</dd>
 <dt><code class="name"><a title="imodels.util.evaluate" href="evaluate/index.html">imodels.util.evaluate</a></code></dt>
 <dd>
 <div class="desc"></div>
@@ -98,6 +102,7 @@ following "Neural Random …</p></div>
 <li><h3><a href="#header-submodules">Sub-modules</a></h3>
 <ul>
 <li><code><a title="imodels.util.convert" href="convert.html">imodels.util.convert</a></code></li>
+<li><code><a title="imodels.util.data_util" href="data_util.html">imodels.util.data_util</a></code></li>
 <li><code><a title="imodels.util.evaluate" href="evaluate/index.html">imodels.util.evaluate</a></code></li>
 <li><code><a title="imodels.util.explain_errors" href="explain_errors.html">imodels.util.explain_errors</a></code></li>
 <li><code><a title="imodels.util.extract" href="extract.html">imodels.util.extract</a></code></li>
