--- experiments/clean.sh
@@ -1,2 +1,0 @@
-rm -f logs/*
-rm -f slurm-scripts/*
