--- experiments/notebooks/matplotlibrc
@@ -1,2 +1,0 @@
-
-figure.dpi: 250
